pyproject.toml
[tool.poetry]
name = "Visual-Slam"
version = "0.1.0"
description = "Simple visual slam."
authors = ["Sacha Hu <hu.sacha.ai@gmail.com>"]
packages = [
    { include = "visual_slam" }
]
readme = "README.md"
maintainers = ["Sacha Hu <hu.sacha.ai@gmail.com>"]
repository = "https://github.com/sachahu1/Visual-Slam"
documentation = "https://www.sachahu.com/docs/visual-slam/latest/index.html"

[tool.poetry.dependencies]
python = ">=3.9,<3.14"
opencv-python = "^4.10.0.84"
open3d-cpu = "^0.18.0"
numpy = ">=1.0,<2"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.5"
pre-commit = "^3.7.1"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^6.2.5"
pytest-mock = "^3.14.0"
coverage = {version = "^6.4.2", extras = ["toml"]}

[tool.poetry.group.documentation]
optional = true

[tool.poetry.group.documentation.dependencies]
Sphinx = "^7.3.7"
sphinxcontrib-napoleon = "^0.7"
sphinx-rtd-theme = "^2.0.0"
myst-parser = "^3.0.1"
sphinx-copybutton = "^0.5.2"
sphinx-autoapi = "^3.2.1"
gitpython = "^3.1.43"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --junit-xml=test_results/tests.xml"
testpaths = ["tests"]

[tool.coverage.run]
source = ["visual_slam"]

[tool.poetry.scripts]
run_visual_slam = "visual_slam.cli:run"
calibrate = "visual_slam.cli:run_calibration"

[tool.ruff]
line-length = 79
indent-width = 2
exclude = [".venv", ".env", ".git", "tests"]


[tool.ruff.lint]
# Add the `line-too-long` rule to the enforced rule set. By default, Ruff omits rules that
# overlap with the use of a formatter, like Black, but we can override this behavior by
# explicitly adding the rule.
extend-select = [
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "ERA", # flake8-eradicate/eradicate
    "I",   # isort
    "N",   # pep8-naming
    "PIE", # flake8-pie
    "PGH", # pygrep
    "RUF", # ruff checks
    "SIM", # flake8-simplify
    "T20", # flake8-print
    "TCH", # flake8-type-checking
    "TID", # flake8-tidy-imports
    "UP",  # pyupgrade
]
fixable = ["ALL"]
ignore = ["SIM117"]

[tool.ruff.lint.isort]
force-single-line = true
lines-after-imports = 2
known-first-party = ["visual_slam"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

